requirements.txt
cos-python-sdk-v5
nacos_sdk_python
pulsar-client==3.1.0